Clean up action plugins, too
