--- plugins/action/net_system.py
@@ -1,18 +1,0 @@
-# (c) 2017, Ansible Inc,
-# GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from __future__ import absolute_import, division, print_function
-
-__metaclass__ = type
-
-from ansible_collections.ansible.netcommon.plugins.action.net_base import (
-    ActionModule as _ActionModule,
-)
-
-
-class ActionModule(_ActionModule):
-    def run(self, tmp=None, task_vars=None):
-        result = super(ActionModule, self).run(tmp, task_vars)
-        del tmp  # tmp no longer has any effect
-        return result
